Entities were edited.AccountNumber and IBAN were edited.And A lot of classes were edited.

+++ src/main/java/org/kodluyoruz/mybank/account/demanddepositaccount/repository/DemandDepositAccountRepository.java
@@ -4,6 +4,6 @@ import org.kodluyoruz.mybank.account.demanddepositaccount.entity.DemandDepositAc
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface DemandDepositAccountRepository extends CrudRepository<DemandDepositAccount, Integer> {
+public interface DemandDepositAccountRepository extends CrudRepository<DemandDepositAccount, Long> {
     DemandDepositAccount findByCustomer_CustomerID(long customerID);
 }
